feat: restructure components, INPRO-2706
